Bind this to _measureLayout in ParallaxImage
Not binding this is an issue when _measureLayout is called as the callback of the onLayout of the View.
In that case, this is the global context so this._container is undefined and the image's size is not updated.
This issue seemed to happen inconsistently, this seems to fix it permanently.